Adding HP and PlanetSpawn logics

+++ Assets/Scripts/Core/IDamagable.cs
@@ -1,4 +1,4 @@
 ﻿internal interface IDamagable
 {
-	void DealDamage ( float damage );
+	void DealDamage ( int damage );
 }
