+++ Assets/Scripts/Core/Weapon/WeaponInfo.cs
@@ -6,7 +6,7 @@ public class WeaponInfo
 {
 	public float Cooldown => _cooldown;
 	public float Weight => _weight;
-	public float Damage => _damage;
+	public int Damage => _damage;
 	public float StartAccelerationDuration => _startAccelerationDuration;
 	public float Acceleration => _acceleration;
 	public bool IsStartAccelerationFinished => _isStartAccelerationFinished;
@@ -16,7 +16,7 @@ public class WeaponInfo
 	[SerializeField] float _acceleration;
 	[SerializeField] float _startAccelerationDuration;
 	[SerializeField] float _weight;
-	[SerializeField] float _damage;
+	[SerializeField] int _damage;
 	[SerializeField] float _cooldown;
 
 	private bool _isStartAccelerationFinished = false;
